+++ consts_big_wasip1.go
@@ -15,6 +15,7 @@ const (
 	supportsLinks         = true
 	supportsNice          = false
 	supportsSymlinks      = false
+	supportsUmask         = true
 )
 
 const userIDSource UserIDSourceType = UserIDSourceIsNone