+++ consts_darwin.go
@@ -17,6 +17,7 @@ const (
 	supportsLinks         = true
 	supportsNice          = true
 	supportsSymlinks      = true
+	supportsUmask         = true
 )
 
 const userIDSource UserIDSourceType = UserIDSourceIsInt