+++ stat.go
@@ -239,6 +239,11 @@ func SupportsSymlinks() bool {
 	return supportsSymlinks
 }
 
+// SupportsUmask returns true if the Umask() function is supported by the OS.
+func SupportsUmask() bool {
+	return supportsUmask
+}
+
 // UserIDSource returns the source of the user's ID: UserIDSourceIsInt,
 // UserIDSourceIsString, or UserIDSourceIsNone.
 func UserIDSource() UserIDSourceType {